Matching game
As part of my Coursera HTML, JavaScript, CSS specialization I have created a matching game. In this game there are 5 smiley faces on the left side of the screen and 4 smiley faces on right side of the screen. A player is suppose to select the extra smiley face on the left side of the screen and click on it to continue the game or else the game is over if player clicks on the wrong face which in part of the same pattern on the right side of the screen, with each correct selection 5 extra smiley faces appear on the left side of the screen and 4 extra smiley faces appear on the right side of the screen.